requirements.txt
Django==4.1.3
django-embed-video